Move tag and createHelpers into helpers.ts
